move integration test folder out of controller
