+++ integration/test/error/error_test.go
@@ -13,8 +13,8 @@ import (
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/giantswarm/operatorkit/controller/integration/testresource"
-	"github.com/giantswarm/operatorkit/controller/integration/wrapper/drainerconfig"
+	"github.com/giantswarm/operatorkit/integration/testresource"
+	"github.com/giantswarm/operatorkit/integration/wrapper/drainerconfig"
 	"github.com/giantswarm/operatorkit/resource"
 )
