+++ integration/test/statusupdate/status_update_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/giantswarm/microerror"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/giantswarm/operatorkit/controller/integration/wrapper/drainerconfig"
+	"github.com/giantswarm/operatorkit/integration/wrapper/drainerconfig"
 	"github.com/giantswarm/operatorkit/resource"
 )
