+++ integration/wrapper/drainerconfig/wrapper.go
@@ -1,8 +1,10 @@
-package configmap
+package drainerconfig
 
 import (
 	"time"
 
+	"github.com/giantswarm/apiextensions/pkg/apis/core/v1alpha1"
+	"github.com/giantswarm/apiextensions/pkg/clientset/versioned"
 	"github.com/giantswarm/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
@@ -13,13 +15,15 @@ import (
 	"k8s.io/client-go/kubernetes"
 
 	"github.com/giantswarm/operatorkit/controller"
-	"github.com/giantswarm/operatorkit/controller/integration/env"
-	"github.com/giantswarm/operatorkit/controller/integration/testresourceset"
+	"github.com/giantswarm/operatorkit/integration/env"
+	"github.com/giantswarm/operatorkit/integration/testresourceset"
 	"github.com/giantswarm/operatorkit/resource"
 )
 
 type Config struct {
-	Resources []resource.Interface
+	HandlesFunc func(obj interface{}) bool
+	Logger      micrologger.Logger
+	Resources   []resource.Interface
 
 	Name      string
 	Namespace string
@@ -28,17 +32,17 @@ type Config struct {
 type Wrapper struct {
 	controller *controller.Controller
 
+	g8sClient versioned.Interface
 	k8sClient kubernetes.Interface
 }
 
 func New(config Config) (*Wrapper, error) {
 	var err error
 
-	var newLogger micrologger.Logger
-	{
+	if config.Logger == nil {
 		c := micrologger.Config{}
 
-		newLogger, err = micrologger.New(c)
+		config.Logger, err = micrologger.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -47,7 +51,10 @@ func New(config Config) (*Wrapper, error) {
 	var k8sClient *k8sclient.Clients
 	{
 		c := k8sclient.ClientsConfig{
-			Logger: newLogger,
+			SchemeBuilder: k8sclient.SchemeBuilder{
+				v1alpha1.AddToScheme,
+			},
+			Logger: config.Logger,
 
 			KubeConfigPath: env.KubeConfigPath(),
 		}
@@ -61,9 +68,10 @@ func New(config Config) (*Wrapper, error) {
 	var resourceSet *controller.ResourceSet
 	{
 		c := testresourceset.Config{
-			K8sClient: k8sClient.K8sClient(),
-			Logger:    newLogger,
-			Resources: config.Resources,
+			HandlesFunc: config.HandlesFunc,
+			K8sClient:   k8sClient.K8sClient(),
+			Logger:      config.Logger,
+			Resources:   config.Resources,
 
 			ProjectName: config.Name,
 		}
@@ -77,13 +85,14 @@ func New(config Config) (*Wrapper, error) {
 	var newController *controller.Controller
 	{
 		c := controller.Config{
+			CRD:       v1alpha1.NewDrainerConfigCRD(),
 			K8sClient: k8sClient,
-			Logger:    newLogger,
+			Logger:    config.Logger,
 			ResourceSets: []*controller.ResourceSet{
 				resourceSet,
 			},
 			NewRuntimeObjectFunc: func() pkgruntime.Object {
-				return new(corev1.ConfigMap)
+				return new(v1alpha1.DrainerConfig)
 			},
 
 			Name:         config.Name,
@@ -98,6 +107,7 @@ func New(config Config) (*Wrapper, error) {
 
 	wrapper := &Wrapper{
 		controller: newController,
+		g8sClient:  k8sClient.G8sClient(),
 		k8sClient:  k8sClient.K8sClient(),
 	}
